+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v.0.12.1
 - Dataset features can now be easily accessed with the property dataset.features
+- `model.test_estimators` with CV will keep using CV if it's refitting the best estimator.
 
 # v0.12.0
 - Permutation importance and Feature importance are now two different plotting methods.